(Temporary?) fix for #1575: images incorrectly loading @2x versions in iPad's iPhone compatibility mode in 2x zoom
UIScreen's scale property (used to check if we're running on a device with high-res Retina Display) appears as an undocumented surprise on iPads running 3.2; in 2x zoom mode it was returning a scale factor of 2.0, which makes a reasonable amount of sense but ended up causing false positives because we thought we were running on a high-res display, but couldn't then rescale the @2x images to the proper logical display size.
Pulled a check from ImageLoader over into [TiUtils isRetinaDisplay] so we can live in a 1x world.

However... this may possibly break in future with iOS 4.x running on iPad, since the newer selectors on UIImage probably are implemented there. A more thorough check should be added here, and possibly elsewhere in the code where scale might be checked.