Initial attempt at Ti.UI.Clipboard implementation for iOS, following the Titanium Desktop version's API.
setText() works, but the getters currently don't (need to fix the thread stuff). Data untested, but in theory should allow setting images (from image, data, or file-backed blobs) and raw-data blobs, though there may not be proper mapping of MIME types to UTIs for the iOS pasteboard.
Special handling of 'image' and 'url' types to explicitly mark as image or URL, as per the Desktop version.